+++ tests/data/forms/fixes/form.e.ts
@@ -1,10 +1,20 @@
+import { expect } from "chai";
 import { Form } from "../../../../src";
 import FormInterface from "../../../../src/models/FormInterface";
 
-const fields = ["places[]"];
+const fields = [
+  "places[]",
+  "test",
+];
 
 const extra = {
   places: ["a", "b", "c"],
+  test: {
+    testExtraProp: 'test',
+    testExtraFunction: () => ({
+     test: 'test',
+    })
+  }
 };
 
 const values = {
@@ -16,9 +26,19 @@ class NewForm extends Form {
     return {
       onInit(form: FormInterface) {
         form.$("places").clear();
+
+        describe("Check extra data and function:", () => {
+          it("$('test') extra prop should be defined and equal 'test'", () => expect(form.$('test').extra.testExtraProp).to.be.equal('test'));
+          it("$('test') extra function should not be undefined", () => expect(form.$('test').extra.testExtraFunction).to.not.be.undefined);
+          it("$('test') extra function should return 'test'", () => expect(form.$('test').extra.testExtraFunction()).to.be.deep.equal({
+            test: 'test'
+          }));
+        });
       },
     };
   }
 }
 
-export default new NewForm({ fields, values, extra }, { name: "Fixes-E" });
+export default new NewForm({ fields, values, extra }, { options: {
+  removeNullishValuesInArrays: true,
+}, name: "Fixes-E" });