+++ tests/data/forms/flat/form.n.ts
@@ -5,6 +5,7 @@ import { checkUser } from "../../extension/vjf";
 import svk from "../../../../src/validators/SVK";
 import vjf from "../../../../src/validators/VJF";
 import FormInterface from "../../../../src/models/FormInterface";
+import OptionsModel from "../../../../src/models/OptionsModel";
 
 const fields = {
   username: {
@@ -39,6 +40,11 @@ const plugins = {
   }),
 };
 
+const options: OptionsModel = {
+  validateOnReset: true,
+  validateOnClear: true,
+};
+
 class NewForm extends Form {
   hooks() {
     return {
@@ -51,4 +57,4 @@ class NewForm extends Form {
   }
 }
 
-export default new NewForm({ fields }, { plugins, name: "Flat-N" });
+export default new NewForm({ fields }, { options, plugins, name: "Flat-N" });